refactor: SessionItem 함수 제거
